+++ app/src/main/kotlin/com/auramusic/app/utils/DiscordRPC.kt
@@ -48,7 +48,7 @@ class DiscordRPC(
             smallText = song.artists.firstOrNull()?.name,
             buttons = listOf(
                 "Listen on YouTube Music" to "https://music.youtube.com/watch?v=${song.song.id}",
-                "Visit AuraMusic" to "https://github.com/chila254/AuraMusic"
+                "Visit AuraMusic" to "https://github.com/TeamAuraMusic/AuraMusic"
             ),
             type = Type.LISTENING,
             statusDisplayType = if (useDetails) StatusDisplayType.DETAILS else StatusDisplayType.STATE,